Render a Rd object out to a markdown file

Render a Rd object out to a markdown file

Usage

rd2markdown(x, fragments = c(), ...)

# S3 method for Rd
rd2markdown(x, fragments = c(), ...)

# S3 method for list
rd2markdown(x, fragments = c(), ...)

# S3 method for `NULL`
rd2markdown(x, fragments = c(), ...)

# S3 method for name
rd2markdown(x, fragments = c(), ...)

# S3 method for alias
rd2markdown(x, fragments = c(), ...)

# S3 method for keyword
rd2markdown(x, fragments = c(), ...)

# S3 method for USERMACRO
rd2markdown(x, fragments = c(), ...)

# S3 method for code
rd2markdown(x, fragments = c(), ...)

# S3 method for RCODE
rd2markdown(x, fragments = c(), ...)

# S3 method for verb
rd2markdown(x, fragments = c(), ...)

# S3 method for pkg
rd2markdown(x, fragments = c(), ...)

# S3 method for default
rd2markdown(x, fragments = c(), ...)

# S3 method for title
rd2markdown(x, fragments = c(), ...)

# S3 method for description
rd2markdown(x, fragments = c(), ..., title = NULL)

# S3 method for author
rd2markdown(x, fragments = c(), ...)

# S3 method for format
rd2markdown(x, fragments = c(), ...)

# S3 method for describe
rd2markdown(x, fragments = c(), ...)

# S3 method for details
rd2markdown(x, fragments = c(), ...)

# S3 method for note
rd2markdown(x, fragments = c(), ...)

# S3 method for source
rd2markdown(x, fragments = c(), ...)

# S3 method for value
rd2markdown(x, fragments = c(), ...)

# S3 method for section
rd2markdown(x, fragments = c(), ...)

# S3 method for examples
rd2markdown(x, fragments = c(), ...)

# S3 method for usage
rd2markdown(x, fragments = c(), ..., title = NULL)

# S3 method for references
rd2markdown(x, fragments = c(), ...)

# S3 method for seealso
rd2markdown(x, fragments = c(), ...)

# S3 method for arguments
rd2markdown(x, fragments = c(), ...)

# S3 method for dots
rd2markdown(x, fragments = c(), ...)

# S3 method for TEXT
rd2markdown(x, fragments = c(), ...)

# S3 method for COMMENT
rd2markdown(x, fragments = c(), ...)

# S3 method for character
rd2markdown(
  x = NULL,
  fragments = c(),
  ...,
  topic = NULL,
  package = NULL,
  file = NULL,
  envir = parent.frame()
)

# S3 method for emph
rd2markdown(x, fragments = c(), ...)

# S3 method for item
rd2markdown(x, fragments = c(), ...)

# S3 method for enumerate
rd2markdown(x, fragments = c(), ...)

# S3 method for itemize
rd2markdown(x, fragments = c(), ...)

# S3 method for LIST
rd2markdown(x, fragments = c(), ...)

# S3 method for tabular
rd2markdown(x, fragments = c(), ...)

# S3 method for tab
rd2markdown(x, fragments = c(), ...)

# S3 method for cr
rd2markdown(x, fragments = c(), ...)

Arguments

x

An Rd object, list, help topic or Rd tag to convert into markdown-formatted help.

fragments

An optional vector of fragment tag names (such as "description", "details" or "title") to filter the Rd object tags.

...

Additional arguments used by specific methods.

title

optional section title

topic

An optional help topic name

package

An optional package name in which to search for the help topic

file

An optional Rd file path from which to load the help topic

envir

An optional environment in which to search for the help topic

Details

rd2markdown is the core function of the package. can work in various types. It accepts .Rd objects extracted with get_rd option but also can directly fetch documentation topics based on the file path or the package. Due to the design, it follows, rd2markdown function has high flexibility. It is worth pointing out that the actual output of rd2markdown function is a character vector of length one that uses special signs to format the document. Use writeLines function to create an actual markdown file using this output. The rd2markdown function parses documentation object based on an innovative idea of treating each .Rd tag as a separate class and implementing methods for those classes. Thanks to that the package is easily extensible and clear to understand. To see the list of currently supported tags please see the Usage section.

See also

Examples

# Using get_rd
rd_file_example <- system.file("examples", "rd_file_sample.Rd", package = "rd2markdown")
rd <- rd2markdown::get_rd(file = rd_file_example)
cat(rd2markdown::rd2markdown(rd))
#> # Rd sampler title
#> 
#> ```r
#> rd_sampler(x, y = TRUE, ...)
#> ```
#> 
#> ## Arguments
#> 
#> - **x**: Rd sampler param
#> 
#> - **y**: Rd sampler param with default
#> 
#> - **...**: Rd sampler ellipsis param
#> 
#> ## Returns
#> 
#> Rd sampler return 
#> 
#> 
#> Rd sampler description with , `Rd sampler in-line code`. And Rd dynamic content, **italics text**, **emphasis text**. 
#> ## Details
#> 
#> Rd sampler details Rd sampler enumerated list 
#> 
#> 1. One
#> 
#> 2. Two
#> 
#> 3. Three
#> 
#> Rd sampler itemized list 
#> 
#>  * One
#> 
#>  * Two
#> 
#>  * Three
#> 
#> 
#> 
#> ||||
#> |--:|--:|--:|
#> |Rd|Sampler|Table|
#> |rd|sampler|table|
#> |character(0)|
#> 
#> ## Note
#> 
#> Rd sampler note 
#> 
#> ## Rd sampler subsection
#> 
#> Rd sampler subsection text 
#> 
#> ## Examples
#> 
#> ```r
#> rd_sampler()
#> ```
#> 
#> ## References
#> 
#> R.D. Sampler. (2021) 
#> 
#> ## See Also
#> 
#> base::print 
#> 
#> ## Author(s)
#> 
#> R.D. Sampler. 
#> 
#> rd_samplers 

# Limit to particular sections
rd <- rd2markdown::get_rd(file = rd_file_example)
cat(rd2markdown::rd2markdown(rd, fragments = c("description", "usage")))
#> 
#> ```r
#> rd_sampler(x, y = TRUE, ...)
#> ```
#> 
#> 
#> Rd sampler description with , `Rd sampler in-line code`. And Rd dynamic content, **italics text**, **emphasis text**. 

# Use file parameter
cat(rd2markdown::rd2markdown(file = rd_file_example))
#> # Rd sampler title
#> 
#> ```r
#> rd_sampler(x, y = TRUE, ...)
#> ```
#> 
#> ## Arguments
#> 
#> - **x**: Rd sampler param
#> 
#> - **y**: Rd sampler param with default
#> 
#> - **...**: Rd sampler ellipsis param
#> 
#> ## Returns
#> 
#> Rd sampler return 
#> 
#> 
#> Rd sampler description with , `Rd sampler in-line code`. And Rd dynamic content, **italics text**, **emphasis text**. 
#> ## Details
#> 
#> Rd sampler details Rd sampler enumerated list 
#> 
#> 1. One
#> 
#> 2. Two
#> 
#> 3. Three
#> 
#> Rd sampler itemized list 
#> 
#>  * One
#> 
#>  * Two
#> 
#>  * Three
#> 
#> 
#> 
#> ||||
#> |--:|--:|--:|
#> |Rd|Sampler|Table|
#> |rd|sampler|table|
#> |character(0)|
#> 
#> ## Note
#> 
#> Rd sampler note 
#> 
#> ## Rd sampler subsection
#> 
#> Rd sampler subsection text 
#> 
#> ## Examples
#> 
#> ```r
#> rd_sampler()
#> ```
#> 
#> ## References
#> 
#> R.D. Sampler. (2021) 
#> 
#> ## See Also
#> 
#> base::print 
#> 
#> ## Author(s)
#> 
#> R.D. Sampler. 
#> 
#> rd_samplers 

# Use topic and package parameters

cat(
  rd2markdown::rd2markdown(topic = "rnorm", package = "stats", 
  fragments = c("description", "usage"))
 )
#> 
#>  Density, distribution function, quantile function and random  generation for the normal distribution with mean equal to `‘mean’` and standard deviation equal to `‘sd’`. 
#> 
#> ```r
#> dnorm(x, mean = 0, sd = 1, log = FALSE)
#> pnorm(q, mean = 0, sd = 1, lower.tail = TRUE, log.p = FALSE)
#> qnorm(p, mean = 0, sd = 1, lower.tail = TRUE, log.p = FALSE)
#> rnorm(n, mean = 0, sd = 1)
#> ```
#>